--- duckchat/duck-ai-internal-api/build.gradle
@@ -1,14 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-}
-
-apply from: "$rootProject.projectDir/gradle/android-library.gradle"
-
-android {
-    namespace 'com.duckduckgo.duckchat.internalapi'
-}
-
-dependencies {
-    // intentionally empty — interface module only
-}
