--- duckchat/duck-ai-store-internal/build.gradle
@@ -1,43 +1,0 @@
-/*
- * Copyright (c) 2026 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'com.squareup.anvil'
-}
-
-apply from: "$rootProject.projectDir/gradle/android-library.gradle"
-
-android {
-    namespace 'com.duckduckgo.duckchat.store.internal'
-    anvil {
-        generateDaggerFactories = true
-    }
-}
-
-dependencies {
-    anvil project(path: ':anvil-compiler')
-    implementation project(path: ':anvil-annotations')
-    implementation project(path: ':di')
-    implementation project(path: ':duck-ai-internal-api')
-    implementation project(path: ':duck-ai-store-impl')
-    implementation project(path: ':duckchat-api')
-
-    implementation Google.dagger
-    implementation "org.nanohttpd:nanohttpd:2.3.1"
-    implementation "com.squareup.logcat:logcat:_"
-}
