+++ duckchat/duckchat-impl/build.gradle
@@ -46,7 +46,7 @@ dependencies {
     implementation project(':sync-api')
     implementation project(':data-store-api')
     implementation project(':cookies-api')
-    implementation project(':duck-ai-store-api')
+    implementation project(':duckchat-store')
 
     anvil project(path: ':anvil-compiler')
     implementation project(path: ':anvil-annotations')
