+++ duckchat/duckchat-impl/src/test/kotlin/com/duckduckgo/duckchat/impl/clearing/DelegatingDuckChatDeleterTest.kt
@@ -18,7 +18,7 @@ package com.duckduckgo.duckchat.impl.clearing
 
 import com.duckduckgo.common.test.CoroutineTestRule
 import com.duckduckgo.duckchat.impl.pixel.DuckChatPixels
-import com.duckduckgo.duckchat.store.api.DuckAiChatStore
+import com.duckduckgo.duckchat.store.impl.DuckAiChatStore
 import dagger.Lazy
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertFalse
