+++ duckchat/duckchat-internal/build.gradle
@@ -36,10 +36,10 @@ android {
 dependencies {
     anvil project(path: ':anvil-compiler')
     implementation project(path: ':anvil-annotations')
-    implementation project(path: ':duck-ai-internal-api')
     implementation project(path: ':internal-features-api')
     implementation project(path: ':duckchat-api')
     implementation project(path: ':duckchat-impl')
+    implementation project(path: ':duckchat-store')
     implementation project(path: ':content-scope-scripts-api')
     implementation project(path: ':content-scope-scripts-impl')
     implementation project(path: ':di')
@@ -54,6 +54,8 @@ dependencies {
     implementation AndroidX.lifecycle.runtime.ktx
     implementation Google.android.material
     implementation Google.dagger
+    implementation "org.nanohttpd:nanohttpd:2.3.1"
+    implementation "com.squareup.logcat:logcat:_"
 
     testImplementation Testing.junit4
     testImplementation Testing.robolectric
