+++ duckchat/duckchat-store/src/main/java/com/duckduckgo/duckchat/store/impl/store/DuckAiBridgeFileMetaEntity.kt
